scantailor: Implement progress notification for configuration generation
To get a look into the ScanTailor 'blackbox', we use the 'psutil' module
to get a list of files opened by the ScanTailor process every 100ms.
With the knowledge of how many images there are and which steps are run,
we can calculate how far into the process we are.