Fix setting of data dir for *nix environments (fixes #117)
Previously the data directory was retrieved from the XDG_CONFIG_DIRS environment variable, which can contain multiple directories. Due to a misinterpretation of the freedesktop.org spec the code assumed that it contained only a single directory, which is why it failed with multiple. We now use XDG_CONFIG_HOME which per the spec can only contain a single directory