gphoto2driver: Remove all configuration options
The decision to do this is based on the fact that libgphoto2 supports an
extremely wide range of cameras, the configuration possibilities for
which vary drastically. This is especially bad between vendors, where
one and the same functionality can sometimes have up to three or four
different names and possible values. Since the gphoto2-cffi bindings
allow changes to the camera configuration through the camera's own UI
during runtime, it was decided to leave it to the user to manually
adjust the camera settings himself. This simplifies the spreads
interfaces significantly, cuts down on moving parts inside the
application where things can go wrong and keeps ugly workarounds and
potentially unstable dependencies on libgphoto2 internals out.